+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 ### Changed
 - `powpow.GrepResult` is now actually importable from the top-level namespace.
+- `powpow.grep` will no longer accept an empty `pattern` string, raising
+  `ValueError` instead.
 
 ## [0.0.3]
 ### Added